Change default ZUUL reference.
The default Zuul reference only gets used when job is invoked directly from
jenkins and not from zuul. So in this case we should provide a reference that
won't be found in Zuul repo. This way we default to checking out master branch from
gerrit repo, not the master branch from zuul repo (unreliable since it may refer
to un-merged changes).

Change-Id: I591f7d8bc744030013e9fb141042ea60fc53dddf