Specify hbase tmp dir
By default the hbase tmp dir is based on the java.io.tmpdir, which by
default is /tmp. The tmpwatch package sets up cronjobs to clean
up unused file in /tmp. If no jobs run for a while and the hbase.version
file gets removed, then hbase gets error on start-up.

So here we specify a new directory that should not be cleaned.

Change-Id: I8282fee7036a4a7fef998def0ce419d4d625c5c8