Provide better build result when no code changes found
Instead of failing the daily build early on, go ahead and run
the build, and check at the last step, before staging (posting)
the built files.

This allows the build to exit successfully. Actually it will probably
be marked UNSTABLE, since the publishing step will fail.

Also make the output of the check a bit more noticeable when no changes
are detected.

Finally change the time of the daily build from 02:30 UTC to 07:30 UTC.
US Pacific time this is a change from 7:30 PM to just after midnight.
This should help the dates make more sense for left coast.

Change-Id: I2a170d1a97250a8c87966bd9fcc4e4bfc673578f