Fix problem with using latest version of python-novaclient v2
Fix problem of using the latest version of python-novaclient
on the Puppetmaster.  The latest version is version 2.18.0 which
is incompatible with the HP Cloud version 13.5 infrastructure.
This fix will pin the version of python-novaclient to 2.17.
Partially undo change I3ff58a6fe70bcac47ec94351e7fc56a75aefc4d2
which pinned python-novaclient to 2.18.0 instead.

Change-Id: Ibe9be3d7df1c28c8b44061804ac352634e0d01aa
Closes-Bug: #1341787