Create check-build job
https://blueprints.launchpad.net/trafodion/+spec/infra-build-verification

This establishes a post-build check script, and a new jenkins job for
core.  Once Jenkins has the new job established, zuul can add it to pipelines.

Also once the check script is propagated to the test slaves, it could also be added
to the existing dcs build job.

Changes to traf-functions.sh so that we don't pollute the environment with
extraneous files before we build.

Change-Id: Idd1e683d35a0dc0e64687fe51fc43c1ce9da84c2