Use absolute directory when looking for core files
Change 203 added a call to find-corefiles.pl using the parameter supplied to
run-traf-regress.sh.  Because that is a relative directory, and the shell
was not located appropriately, the result of "cd $DIR" was this error:

/usr/local/bin/run-traf-regress.sh: line 99: cd: ./trafodion/core: No such file or directory

It was harmless, in that the regression script still ended with success.

We now cd to one level above $MY_SQROOT, then use $PWD as the parameter
for find-corefiles.pl.

Change-Id: Ib0f403ecf252f04dcf0c5ecb9492c99d6abe33a2