Added optional argument to report_on_corefiles
Function report_on_corefiles was merged into the master branch today
as change 208.  Comparing phoenix_quick output before and after that
merge, the extra output is:

  +  report_on_corefiles
  + [[ -d '' ]]

The problem was that the run-phoenix_test.sh does not use SQ_MYROOT.
The solution is to add a parameter to report_on_corefiles for the top
directory within which core files are found.

The core-regress-core jobs correctly look for core files because
MY_SQROOT is used in run-traf-regress.sh.  No change is needed
there because when report_on_corefile has no arguent, it now uses
the directory above $MY_SQROOT.

Change-Id: Id71e64c478627dc63b659ecc28de331079cee0ef