Core file detection fix
Core file detection on slaves did not work, as recorded in
bug 1359881 because the core file names use the short hostname,
not the long form.  For example, one file was

  core.slave07.21791.sqlci

whereas this script expected

  core.slave07.trafodion.org.21791.sqlci

This has been corrected, the script uses whatever is returned
by "hostname", and not "hostname -f".

Change-Id: Icbe0ac852c50524b05b1c2a437dcc7403febbc04