Change recheck/reverify keywords requirements for ease of use
This relaxs the comment format for recheck requests.

The end of the regex is not specified, so it is open-ended to provide
either the bug number or some comment of whay a recheck is needed.

The bug number will still be used by the recheck tracking feature,
and will be required for gate.

Either keyword can be used interchangeably (because which to use was
confusing to developers). If the Verified attribute is -1 or 0, it
indicates that we don't have a valid check test, so we enter the check
queue. If the attribute is 1, 2, or -2, that indicates the check test
has passed previously, so we enter the gate queue.

That scheme may have a slight hole, since the gate queue has other
requirements, particularly Workflow=1 (approved). So if the Verfied
attribute is -2 and Workflow is 0, the keyword would not trigger anything.
But such scenarios are odd cases, since it would mean that it was approved,
but then approval was taken away. In those odd cases, testing has to
be triggered in other ways, such as a new patch set.

This change is not fully tested, but only affects the recheck comment
functionality.

Change-Id: Iafe8fbb0dc188717e3fda56d66f181a09ec0cb6b
Closes-Bug: 1340320