Update standard puppet modules to new mysql and add cloudera
This updates all the modules in use for new mysql puppet module.
Includes porting install_modules feature from openstack-infra
to install puppet modules from git URL. We use this to pull
in a forked version of puppet-dashboard. The forked version
was necessary to also make it compatible with new mysql module.

All of the mysql updates allows us to use the cloudera module.

Change-Id: I55b52d24af909113e98c1ba9b34f19a7ed3255c8