Check for absolute filenames used in dll references
Added a new check of the workspace after a build of
trafodion/core.  This uses file sqf/build-scripts/find-abs-dlls
which was added to trafodion/core in change 392 on 2014-09-15,
and should be in all open and new changes to trafodion/core.

This works only in trafodion/core because it uses a script there.

Two other "success" messages were added here, so a successful
run of check-git-status.sh is now reported this way:

  Success: Build modified 0 versioned files
  Success: Build created 0 untracked files
  Success: Build created 0 dll references to absolute filenames

Change-Id: Ie0230ed4a871294ec27c2fc04c231fa795c3e6ed