Fix backtrace utility for long core file names
The core_bt script did not work when the core file name was too long.
For example:

  $ core_bt
  core file  : -rw-------. 1 jenkins jenkins 407957504 Sep 23 21:58
                       ./sqf/rundir/seabase/core.slave01.10457.sqlci
  sh: -c: line 0: unexpected EOF while looking for matching `''
  sh: -c: line 2: syntax error: unexpected end of file
  . . .

The problem was incomplete output from "file".  For example,

$ file ./sqf/rundir/seabase/core.slave01.10457.sqlci
./sqf/rundir/seabase/core.slave01.10457.sqlci: ELF 64-bit LSB core file
 x86-64, version 1 (SYSV), SVR4-style, from
 '/home/jenkins/workspace/core-regress-seabase-cdh4.4/
 trafodion/core/sqf/export/b'

The correct last part of the executable's file name is
 trafodion/core/sqf/export/bin64/sqlci

Now this script checks the executable name better, and if it's not
found in output from "file", tries the last part of the core file name,
e.g., extracts "sqlci" from "core.slave01.10457.sqlci".  Then "which"
is used to find the executable location.  This works well only when
the script's PATH is the same as when the executable ran.

Change-Id: I43dedd73c83d35fa8718aeb3f254315f501d73e7