Fix DCS argument for core-regress on cm distro
In the case of sql dev regressions, we were passing in empty
argument for DCS server count. That put -1 into servers file,
which made DCS fail.

Not sure why that caused HBase access problem. But the intent
is to start DCS even in the case we may not utilize it.

Change-Id: Ie09f52158c818aa1eccfda1947feb68d596da8cf