Create HDFS directory for hive suite
In the "manual" installed distros, the /users/hive directory is
chown'd to jenkins user, so the regressions have permission to
create the intermediate directory.  In this case, we won't cheat
so much, but create the intermediate directory and give ownership
to trafodion user, who runs the tests.

We need to do it after installer runs, since installer creates the
trafodion user.

Change-Id: I0a49e2fe14577d7ae614bd8adb1cf8040676568b