Fix problem with duplicate declaration of some packages
Fix problem with duplicate declaration of python-devel/python-dev
and libffi-devel/libffi-dev.  The python-devel/python-dev package
has been removed since it is already installed by the pip module.
The libffi-devel/libffi-dev will be installed as part of the base
package if the machine is not a jenkins slave.  If the machine is
a jenkins slave it will be installed by the jenkins slave module.

Change-Id: Ie69c1d8408153c92d027442ad4b972d9466c7f6c
Implements: blueprint infra-data-back-up