Fix Python and Openstack CLI on slaves
Pinned Python tox package defined in the Jenkins slave manifest to
1.7.2 which is the version expected by the traf::python27 class.
Split Openstack CLI installation by moving the installation of the
CLI on Jenkins slaves to the traf::python27 manifest. This is because
some of the Openstack CLI programs require Python 2.7 which is not
the default on Jenkins slaves. Also all the Openstack CLI programs
have been pinned since the latest version does not work with the
HP Cloud environment.

Change-Id: Ib145bc2f0603e1864eee9d2d0a08cf4e13cfc67d
Implements: blueprint infra-data-back-up