Update run-phoenix script to set more environment variables
Currently the phoenix_test.py script assumes that the config
directory for Hadoop, HBase, and Hive are set in a particular
location. This change will set the environment variables
HBASE_CNF_DIR, HIVE_CNF_DIR, and HADOOP_CNF_DIR so that the
phoenix_test.py will pick up those config directories from the
environment instead. This will not take effect until a future
check-in in the phoenix_test repository enables this.

Change-Id: I00f88e62929876e17d573466defae1b9feb40976