Improvements to scripts for gate-core-check-for-binary
The scripts which check a source tree for binary files and bad
source were creating a temp file in the top level of the source
files being examined.  This was an unnecessary write into a
directory which only needs to be read.  When a non-owner ran
the scripts, they failed with "Permission denied".

For both scripts, this change uses /tmp for a temporary file,
not the directory being examined.

Change-Id: I034cb3ab89e6160691a0b0199a2b8b59aabe9895