Use new script to catch unresolved merge conflicts
Added new builder "assert-no-bad-source-files" which uses
find-bad-source.sh to scan for merge conflicts which
should not be in source, such as:

  =======
  <<<<<<< HEAD
  grant component privilege "CREATE" on SQL_OPERATIONS to "PUBLIC";
  get privileges on component sql_operations for "PUBLIC";
  =======
  >>>>>>> bff2742... ANSI Schema

This change can be done now, because the script added in change 852
is now available on jenkins02 as /usr/local/bin/find-bad-source.sh.

This replaces changes 864 and 867 which have been abandoned.

Change-Id: I90cd68af72012226ae0f37d77bd383f3fc240d7c