Update vim configuration for Python and Puppet
Vim configuration has been moved to the following GitHub
project : https://github.com/trafodion/vim-setup.git
This is due to the fact that our Vim configuration is
coming from many different sources. This way we will be
able to keep track of all the sources.

This change only removes the directory /root/.vim. A
following change will recreate it and link it to the
GitHub repository.

Change-Id: Id40961383cfac49d986940ea5c78baea31fcdcdc