Recognize .psd files as binary files
Change 956 added a new binary file, which "file" reports as:

  src/main/resources/dcs-webapps/master/datatables/images/Sorting icons.psd :
     Adobe Photoshop Image, 19 x 19, RGBA, 4x 8-bit channels

The word "Image" was not recognized by the script looking for binary files,
thus https://jenkins02.trafodion.org/job/gate-dcs-check-for-binary/37/
failed with:
    ERROR: Unknown file type not allowed

This change adds "Image" to the strings recognized in "file" output.

The new file name "Sorting icons.psd" contains a space, which is not in any
existing files in trafodion/core, trafodion/dcs and trafodion/infra.  We
should be alert for the space causing problems in other scripts.

Change-Id: I5b579e7f6ffd0d887dafbe9846f2043a7e44a428