Quiet down the cron emails
The cron jobs set up on some of the infrastructure machines are
sending out way too many emails regardless of whether there is a
failure in the cron job or not. The open-source shell script cronic
will serve as a wrapper for cron jobs so only errors are reported.
More information about cronic can be found @ http://habilis.net/cronic/.
Change was also make to the purgeDashboardDB.sh to suppress the
very long list of Ruby Gem warning the scripts was generating on every
run.

Change-Id: I4b952fb31fddc60908cd5a79cb7f523d4d72022c