Add options to skip starting and stopping Trafodion in test scripts
Add options to pass in a parameter so the script will skip starting
and/or stopping Trafodion. This will make it easier when debugging
issues found by the tests launched by these scripts.  The options
are SKIPSTART, SKIPSTOP, SKIPSTARTSTOP. To use one of these
options it needs to be the first input parameter to the script. A
lso add core file reporting to the JDBC tests.

Change-Id: I3461d99a7500f1c0e0a71fde94108ddb558201fd