Enable use of Object Storage for builds
This enables our builds to be able to publish our builds to
Object Storage/CDN/Swift.  The file useObjectStorage.sh simplifies
interaction with Object Storage. It also contains the logic from
the backupToObjectStorage.sh file which will be obsoleted later.
Add Jenkins publisher traf-upload-to-cdn which will invoke
useObjectStorage.sh to upload builds to Object Storage. The
publisher traf-upload-to-cdn will now be invoked for traf-pub-*
and installer-pub jobs. A following review will enable the
downloads site to point to the files in Object Storage.

Implements : blueprint infra-download-cdn

Change-Id: I9ed99c13b5c89127f141faf52aa80fb8fec63c8c