Inital set up of a puppet master for Puppet 3.x
pserver variable is set in site.pp so that we can roll nodes or
node-types one by one to use 3.x.

Deleted a good bit of unused stuff so that we don't have to maintain
changes for 3.x. If we do need those modules, we'll want newer versions
anyway.  More clean-up could be done. This is a first cut.

Change-Id: I4711a51b9b7922b784489ee6dd0c2d5102945ac2