Enable fetching files from CDN to enhance speed of download site
A previous commit enabled the automatic upload of our builds to
the CDN. This commit enables the fetching of our files from the
CDN which will greatly speed up downloads from our site. The
index page has been re-written to use REST API to interact with
the CDN. A new PHP library httpful.phar is needed for this to work.

Change-Id: I13cbdd7125f97ad50c0cf080f568f2a1213f3aef
Implements: blueprint infra-download-cdn