Fix problem with publishing .tar.gz files to the CDN
Changed the algorithm so that files are individually uploaded.
This way they can be processed correctly. Apparently, Swift
also has a bug where you specifying the header content-type is
ignored during upload. To get around that problem we run a POST
command after uploading to make sure the headers we want are set.
Also changed /usr/local/binuseObjectStorage.sh to not be world
readable/writeable since we have user information in that file.
Due to that change jenkins needs to be given sudo access to that file.

Change-Id: If6ae68f306f28acb7c0d20f2201bfb8727dc26f2