Check-node script to record disk space
This will be used to track disk usage in case we have repeat
of failure mode where all root filesystem space is consumed.

Disk space usage will be output to the stdout (job console)
and also to logs in jenkins home directory.

So far this script is not used. Need to let it propagate
out to all test slave nodes before updating jenkins to
use it in various jobs.

Partial-Bug: #1430876

Change-Id: Ie4bf356d13930456a94f36ad4a4983d0e446ea3a