Include check-node step in test jobs
Add check-node at beginning and end of each test job, but not
build or other quick jobs.

Implement end of job check as a post-build publisher. This will
run the script even at the end of jobs that time-out.

Also add path of HDFS data storage for HW configs.

Partial-Bug: 1430876

Change-Id: I12d64b61ead6b3a7bb97642b4ed14fdba788d05e