Move execution of cluster set-up scripts out of puppet
* Cluster set-up
Moving it out of puppet automation allows it to be run when initially
setting up a VM and from within test jobs, rather than asynchronously
from puppet.

Merge hbase-clean.sh into respective cluster set-up scripts.
amcluster.sh for AHW and cmgr.sh for CM.

Although it seems we are adding a lot to the test jobs, most of
the cluster_setup script runs really fast if the cluster configuration
is already in good shape. The combined script seems a bit more efficient
and so we get through that portion of the test job about 5 minutes
faster than the current hbase_cleanup process.

* Remove support for old-style installer
Remove execution of multi-script installer. It is
no longer needed and simplifies the usage of hbase cleanup.

* Clean up old config files
Delete hadoop configuration files that were used in the CDH and HDP
configurations that did not use cluster manager nor trafodion installer.

* TPC-DS data set-up
Remove tpcds puppet module. This function will now be done via the
install_hadoop_regr_test_env script in core repo. It is only run
if the hive test suite is being run and only if the store_orc table
does not already exist. In future, we can remove that check if a
check is added in the core script. We don't want to run the script
every time since it is redundant and takes 10-15 minutes to run on
a test VM.

Tested puppet and various "node-test" jenkins jobs on both AHW2.2 and
CM5.3 nodes. Also tested generating TPC-DS data in hive regression job
on AHW2.2 node.

Change-Id: I4f301efdde1d6e3265b2d148f1802518f8bf80fb