Make corefile backtraces available on client-side tests
For client-side tests that are running as jenkins user, there may
be corefiles owned and readable only by jenkins or by trafodion.

So we call the reporting function with -u option, which now makes all
corefiles owned by the calling ID (jenkins) world-readable and then
call the backtrace command as the alternate ID (trafodion) which should
now be able to read both kinds of corefiles.

Change-Id: Ib26f56683a32bd3a4a13e1958e2e944bd6efc62c
Closes-Bug: 1436480