Fix ignored-files check
The ignored-files check was not returning the proper error code, because
the source_env function was setting "rc" variable. The check-git-status.sh
was using the same variable for different purpose.  So, make rc local to the
function.

This fix could not be made until the git ignore issues were cleaned up, but
that fix is now merged in core.

Change-Id: I5fa3de3d3c8962f481988e215bf380a0ee127ccb