Fix zuul status reporting (again)
Previous change to home.css prevented the background color from
changing on alternate rows in zuul pipeline, but it still did
not show the connector line between a even number row and the
the following one.  So we must explicitly define background
for td elements with line class and in even number rows.

This glitch only shows up when there are three or more connected
changes, usually in gate pipeline.

Change-Id: I9ce97c23ee0248225dfe597f292bf0e5c9b1a1f4