Remove timeout on node-test job
The timeout-var parameter also did not work as expected. It apparently
exports a variable to a build job, not take a variable from environment
or parameter.

Since variable timeout value is not working, we'll remove it for node-test
jobs altogether. These have to be manually started, so we'll leave it up to
the person to monitor such jobs.

Change-Id: Ib71e8e9b3b94f57c955f8605cb322d1d8b72e9e9