Remove zuul trigger to avoid notification timeout
Specifying zuul trigger very nicely adds various ZUUL_* parameters,
which are handy for running the job manually. It also adds a notification
to beginning and end of job with a bad default URL, which causes notification
timeout. Have not found a proper URL to use, so just remove this feature.

If we want to provide parameters to manually mimic Zuul, we can add those
with parameter directives instead of using this trigger.

Change-Id: I19246a445d2466829054b7b718b3c1fe3a2bc33e