Enhance PyODBC logging
Moved log file copying logic out of run-pyodbc_test.sh to JJB so
if a build fails due to timeout all the necessary files will actually
be copied to the logs server. Also added a command at the end of
run-pyodbc_test.sh to print out a list failing tests.

Change-Id: If0d898018903242a84c9a6d9ad8c2a0f4ee681dd