Make staging (publishing) step part of build-product job
Prior change to publishing jobs did not take into account that
stage-traf.sh script relies on some files left around by git-prep.
So we really need staging to run in the build workspace.

Rather than do redundant builds, this makes staging a step in the
regular build, but it only does something depending on the pipeline
the job runs in. If running a regular check/gate job it will exit
immediatly.

Change-Id: I0ef2858775bfcb16b246dee698d75364e23029b7