Fix static-checks file lists
Previous version listed all file changes of all the commits in the
pull-request branch. So if a later commit changed a file path, it
was still possible for the intermediate commit to cause a failure.

Instead of show command to list every file change, we use diff to
show the final change list of the entire branch, diffing target
branch versus the PR merge branch.